Create Main
Does Not work. Its kind of hard to explain, but basically, it checks a bunch of pixels in row major (i think) order and returns an arraylist of point objects. It then shoots them all at once. So its horribly efficient, if you were to run it, it would hardly ever find the points and if it did, the person would no longer be there anymore anyway. The person I got this idea from put a whole screenshot into Microsoft paint and went from there and I have no idea how to do that. Also, i was experimenting with the code alot, so I removed some stuff so I dont know if it would even run if tested, but the main idea is there and I think the problem can be identified.